% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Antithetic.method.R
\name{Antithetic.method}
\alias{Antithetic.method}
\title{The antithetic variables method for variance reducing}
\usage{
Antithetic.method(sigma, n)
}
\arguments{
\item{sigma}{parameter in Rayleigh distribution function}

\item{n}{the number of observations}
}
\value{
The percent reduction in variance by using antithetic variables method
}
\description{
The antithetic variables method to variance reducing in samples from a Rayleigh distribution
}
\examples{
\dontrun{
Antithetic.method(1, 1000)
}
}
