% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betacdf.R
\name{Betacdf}
\alias{Betacdf}
\title{A function to compute a Monte Carlo estimate of the Beta(a, b) cdf}
\usage{
Betacdf(x, alpha, beta)
}
\arguments{
\item{x}{observations}

\item{alpha}{shape parameter}

\item{beta}{shape parameter}
}
\value{
a Monte Carlo estimate of the Beta(a, b) cdf
}
\description{
A function to compute a Monte Carlo estimate of the Beta(a, b) cdf
}
\examples{
\dontrun{
Betacdf(5,3,3)
}
}
